Create subgoalsgrid as a molecule component
