+++ frontend/src/components/atoms/Navbar.tsx
@@ -7,7 +7,7 @@ const Navbar = () => {
   const history = useHistory();
 
   return (
-    <Flex w="100%" align="center" h="70px" px="8">
+    <Flex w="100%" bg="white" zIndex="999" align="center" h="70px" px="8" position="sticky" top="0">
       <Box>
         <Link fontWeight="bold" color="cyan.600" fontSize="1.5em" as={RouteLink} to="/">
           Trondheim kommune SDG-Ontologi
