+++ frontend/src/components/pages/OntologyPage.tsx
@@ -1,13 +1,18 @@
+import { Stack } from '@chakra-ui/react';
 import React from 'react';
 import Graph from '../atoms/Graph';
 import SearchBar from '../atoms/SearchBar';
 import DetailView from '../molecules/DetailView';
+import SubGoalsGrid from '../molecules/SubGoalsGrid';
 
 const OntologyPage: React.FC = () => (
   <>
     <SearchBar limit={5} margin="2em" />
     <Graph />
-    <DetailView />
+    <Stack spacing="10">
+      <DetailView />
+      <SubGoalsGrid />
+    </Stack>
   </>
 );
 export default OntologyPage;
